# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Decameron API",
        "description": "Documentación de la API de gestión de hoteles y configuraciones de habitación",
        "contact": {
            "name": "Equipo de Desarrollo Decameron",
            "email": "soporte@decameron.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/v1/hoteles": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Listar todos los hoteles",
                "operationId": "9eb12e05feff61ceccef81a04296b6c1",
                "responses": {
                    "200": {
                        "description": "Listado de hoteles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Hotel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Crea un hotel",
                "description": "Crear un nuevo hotel",
                "operationId": "dd3a81d374086615c76b9f80c5afb6bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            },
            "servers": [
                {
                    "url": "http://127.0.0.1:8000"
                }
            ]
        },
        "/api/v1/hoteles/{hotel}/configuraciones": {
            "servers": [
                {
                    "url": "http://127.0.0.1:8000"
                }
            ]
        },
        "/api/v1/hoteles/{id}": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Obtiene un hotel por ID",
                "description": "Obtener detalle de un hotel",
                "operationId": "abb50a02958754cec9e05362220bd02f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Actualiza un hotel",
                "description": "Actualizar un hotel existente",
                "operationId": "8b245b0759a0f66b1076146d9a1456d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Elimina un hotel",
                "description": "Eliminar un hotel",
                "operationId": "02676ba05670bf0ed5752b9ec551eea1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Hotel eliminado"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            }
        },
        "/api/v1/hoteles/{hotelId}/configuraciones": {
            "get": {
                "tags": [
                    "Configuraciones"
                ],
                "summary": "Listar configuraciones de un hotel",
                "operationId": "d150226b8c4ea6b586ccfac0478bb03a",
                "parameters": [
                    {
                        "name": "hotelId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de configuraciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HotelRoomConfig"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Configuraciones"
                ],
                "summary": "Crear configuración de habitación",
                "operationId": "06a9f27366e12539ce2a82c5a0f29bb6",
                "parameters": [
                    {
                        "name": "hotelId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHotelRoomConfigRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Configuración creada"
                    }
                }
            }
        },
        "/api/v1/configuraciones/{id}": {
            "delete": {
                "tags": [
                    "Configuraciones"
                ],
                "summary": "Eliminar configuración",
                "operationId": "3de3cdf194d572a4a1959ba2953078b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Configuración eliminada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreHotelRequest": {
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "nit": {
                        "type": "string"
                    },
                    "direccion": {
                        "type": "string"
                    },
                    "ciudad": {
                        "type": "string"
                    },
                    "max_habitaciones": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "StoreHotelRoomConfigRequest": {
                "properties": {
                    "tipo_habitacion": {
                        "type": "string",
                        "enum": [
                            "ESTANDAR",
                            "JUNIOR",
                            "SUITE"
                        ]
                    },
                    "acomodacion": {
                        "type": "string",
                        "enum": [
                            "SENCILLA",
                            "DOBLE",
                            "TRIPLE",
                            "CUADRUPLE"
                        ]
                    },
                    "cantidad": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Hotel": {
                "properties": {
                    "id": {
                        "description": "ID del hotel",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del hotel",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección",
                        "type": "string"
                    },
                    "ciudad": {
                        "description": "Ciudad",
                        "type": "string"
                    },
                    "nit": {
                        "description": "NIT o registro fiscal",
                        "type": "string"
                    },
                    "max_habitaciones": {
                        "description": "Máximo de habitaciones",
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "HotelRoomConfig": {
                "description": "Configuración de tipo de habitación de un hotel",
                "required": [
                    "id",
                    "hotel_id",
                    "tipo_habitacion",
                    "acomodacion",
                    "cantidad"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hotel_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "tipo_habitacion": {
                        "type": "string",
                        "enum": [
                            "ESTANDAR",
                            "JUNIOR",
                            "SUITE"
                        ]
                    },
                    "acomodacion": {
                        "type": "string",
                        "enum": [
                            "SENCILLA",
                            "DOBLE",
                            "TRIPLE",
                            "CUADRUPLE"
                        ]
                    },
                    "cantidad": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresar: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Hoteles",
            "description": "Endpoints para gestionar hoteles"
        },
        {
            "name": "Configuraciones",
            "description": "Configuraciones"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ]
}